third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-ui/async-call-stack-url-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that call stack sidebar contains correct urls for call frames.

Set timer for test function.
Script execution paused.
f3unformatted-asy…s:formatted:12
setTimeout (async)
f2unformatted-asy…js:formatted:8
setTimeout (async)
testFunctionasync-call-stack-url.js:13
setTimeout (async)
scheduleTestFunctionVM:3
(anonymous)VM:1
Script execution resumed.
